+++ src/time_evolution/time_evolution.jl
@@ -173,6 +173,7 @@ struct TimeEvolutionStochasticSol{
     TS<:AbstractVector,
     TE<:Union{AbstractMatrix,Nothing},
     TEA<:Union{AbstractArray,Nothing},
+    TEM<:Union{AbstractArray,Nothing},
     AlgT<:StochasticDiffEqAlgorithm,
     AT<:Real,
     RT<:Real,
@@ -183,6 +184,7 @@ struct TimeEvolutionStochasticSol{
     expect::TE
     average_expect::TE # Currently just a synonym for `expect`
     runs_expect::TEA
+    measurement::TEM
     converged::Bool
     alg::AlgT
     abstol::AT
@@ -349,7 +351,7 @@ function _stochastic_prob_func(prob, i, repeat, rng, seeds, tlist; kwargs...)
         prob.prob.tspan[1],
         zeros(kwargs[:n_sc_ops]),
         zeros(kwargs[:n_sc_ops]),
-        save_everystep = false,
+        save_everystep = getVal(kwargs[:store_measurement]),
         rng = traj_rng,
     )